update component SINA for qanary 3.1.2 (https://github.com/anbo-de/team-tasks/issues/60)
