+++ pom.xml
@@ -44,7 +44,7 @@
                             <module>qanary-component-REL-RelationLinker2</module>
                             <module>qanary-component-REL-RelNliod</module>
                             <module>qanary_component-CLS-CLSNLIOD</module>
-                            <module>qanary_component-QB-Sina</module>
+                            <module>qanary-component-QB-Sina</module>
                             <module>qanary-component-QE-SparqlExecuter</module>
                             <module>qanary_component-QB-SimpleRealNameOfSuperHero</module>
                             <module>qanary_component-QB-ComicCharacterAlterEgoSimpleDBpediaQueryBuilder</module>
