+++ qanary-component-QB-Sina/src/main/resources/config/application.properties
@@ -1,0 +1,41 @@
+
+# Update the port number
+server.port=8098
+spring.application.name=SINA
+spring.application.description=${spring.application.name} is a Qanary component
+# Update the URL of the qanary pipeline
+spring.boot.admin.url=http://localhost:8080
+spring.boot.admin.client.url=${spring.boot.admin.url}
+# the service url
+#spring.boot.admin.client.service-base-url=http://0.0.0.0:${server.port}/
+spring.boot.admin.client.instance.service-base-url=http://0.0.0.0:${server.port}/
+
+# log level definitions
+# change logging level in production
+logging.file=/var/log/application.log
+logging.level.*=INFO
+logging.level.org.springframework.core:WARN
+logging.level.org.springframework.web:WARN
+logging.level.org.springframework.beans:WARN
+logging.level.org.springframework.boot:WARN
+logging.level.org.apache.catalina.core:WARN
+logging.level.org.springframework.jmx:WARN
+logging.level.org.springframework.mock:WARN
+logging.level.org.springframework.test:WARN
+logging.level.eu.wdaqua.qanary:DEBUG
+
+# show colors in IDE console
+spring.output.ansi.enabled=always 
+
+# location inside of class path
+sina.jarfilelocation=sina-0.0.1.jar
+### SSL configuration
+# the path to the key store that contains the SSL certificate, e.g., classpath:keystore.p12
+server.ssl.key-store=
+# the password used to access the key store
+server.ssl.key-store-password=
+# the type of the key store (JKS or PKCS12)
+server.ssl.key-store-type=
+# toggle whether HTTP or HTTPS should be used (if SSL is set up)
+server.ssl.enabled=false
+
