--- qanary_component-QB-Sina/src/main/java/eu/wdaqua/qanary/sina/NoSinaFileProvidedException.java
@@ -1,10 +1,0 @@
-package eu.wdaqua.qanary.sina;
-
-public class NoSinaFileProvidedException extends RuntimeException {
-
-	private static final long serialVersionUID = 473507761065615377L;
-	
-	public NoSinaFileProvidedException() {
-		super("No SINA JAR file provided as application parameter 'sina.jarfilelocation'.");
-	} 
-}
