--- qanary_component-QB-Sina/src/main/resources/templates/description.html
@@ -1,14 +1,0 @@
-<!DOCTYPE HTML>
-<html xmlns:th="http://www.thymeleaf.org">
-<head> 
-    <title>SINA: Qanary Service Status page</title> 
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-</head>
-<body>
-    <h1>SINA</h1>
-    <p th:text="'Hello World, this is my first Qanary service!'" />
-    <!-- 
-	make a statement about the purpose of your component here
-    //-->
-</body>
-</html>
